presentation/pyproject.toml
[project]
name = "build-detective-presentation"
version = "0.1.0"
description = "Build Detective - AI-powered CI/build failure analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.34.0",
    "pyyaml>=6.0",
    "dataclasses-json>=0.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
build-detective = "src.main:main"
